% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/many_sample_means.R
\name{many_sample_means}
\alias{many_sample_means}
\title{Calculates many sample means of the same vector.}
\usage{
many_sample_means(vec, n, reps)
}
\arguments{
\item{vec}{A vector of numbers}

\item{n}{An integer representing the number of samples to take}

\item{reps}{An integer representing the number of sample means to generate}
}
\value{
A vector of numbers representing the sample means generated
}
\description{
Calculates many sample means of the same vector.
}
