% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_mean.R
\name{sample_mean}
\alias{sample_mean}
\title{Calculates the sample mean of a vector.}
\usage{
sample_mean(vec, n)
}
\arguments{
\item{vec}{A vector of numbers}

\item{n}{An integer representing the number of samples to take}
}
\value{
The sample mean of a vector with n items selected (with replacement)
}
\description{
Calculates the sample mean of a vector.
}
