% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_means_ns.R
\name{sample_means_ns}
\alias{sample_means_ns}
\title{Generates many sample means for each number in ns}
\usage{
sample_means_ns(vec, reps, ns)
}
\arguments{
\item{vec}{A vector of numbers}

\item{reps}{An integer representing the number of sample means to generate for all n in ns}

\item{ns}{A vector of n integers which will be looped trough and used to generate sample means with those n samples}
}
\value{
A tibble showing the given n used to generate each sample_mean (cols: n, column_means)
}
\description{
Generates many sample means for each number in ns
}
